+++ custom_components/z2m_irrigation/switch.py
@@ -8,7 +8,7 @@ from .manager import ValveManager, Valve
 from .const import DOMAIN, MANUFACTURER, MODEL, SIG_NEW_VALVE, sig_update
 
 async def async_setup_entry(hass, entry: ConfigEntry, async_add_entities):
-    mgr: ValveManager = hass.data[DOMAIN][entry.entry_id]
+    mgr: ValveManager = hass.data[DOMAIN][entry.entry_id]["manager"]
     async def _add(v: Valve):
         await async_add_entities([ValveSwitch(mgr, v)], True)
     for v in list(mgr.valves.values()):